% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{alldata}
\alias{alldata}
\title{Price index of SP500 and DAX}
\format{
Data frame with two columns and 6211 rows. The row-names includes the dates.
\describe{
\item{value}{Price index}
\item{unit}{index}
}
}
\usage{
alldata
}
\description{
Prices of stock market for the U.S. SP500 and German DAX.
}
\examples{
  alldata
}
\keyword{datasets}
