% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_information.R
\name{Fisher_Information}
\alias{Fisher_Information}
\title{Compute the Fisher Information (FI) from a given time series.}
\usage{
Fisher_Information(rets, splits = c(3:12))
}
\arguments{
\item{rets}{A numeric vector with returns (or values) to be computed the
Fisher Information.}

\item{splits}{Choose the number of intervals in which you are going to split
all the values.}
}
\value{
A Dataframe with two columns: the used split, ans the  FI measure.
}
\description{
The FI is useful because it serves as a measure that acts as a "detector"
of changes in series dynamics.  Please see the paper by Martin, Perez,
and Plastino.
}
\details{
Reference: Martin, M. T., Perez, J., & Plastino, A. (2001). Fisher
information and nonlinear dynamics. Physica A: Statistical Mechanics and its
Applications, 291(1-4), 523-532.
}
\examples{
# Compute the FI for white noise using splits from 3 to 12.
white_noise <- rnorm(50)
r <- Fisher_Information(rets = white_noise, splits=c(3:12))
library(ggplot2)
ggplot(r, aes(x=factor(split), y=fi)) + geom_boxplot(fill='#A4A4A4',
color="black") + labs(title="Fisher-Information",x="Split window", y = "FI") +
theme_classic()

}
