% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logvariations_computation.R
\name{logvariations_computation}
\alias{logvariations_computation}
\title{Computes the variations (log returns) of L periods.}
\usage{
logvariations_computation(prices, P, overlap)
}
\arguments{
\item{prices}{A vector with prices or values.}

\item{P}{A period of the variation or return. If for example L=5, the variations are weekly in a finance domain.}

\item{overlap}{A number with the length of the overlap. The overlap has to be grater than 0 and less or equal than L. If overlap is L, then there is no overlapping. If overlap is 1, then the starting value to compute the next variation will be p(t+1).}
}
\value{
A vector value with log variations.
}
\description{
logvariations_computation() Computes the variations of values of a time series {p} as log p(t+L) - log p(t).
}
\examples{
# Compute the log return of the SP500 of 3 days, with overlap of 3 days (no overlapping)
logvariations_computation(prices = alldata$`^GSPC`, P=3, overlap=3)
\dontrun{
# overlap > P, that's not possible (gives error)
logvariations_computation(prices = c(100, 101, 99, 100), P=1, overlap=3)
# overlap = 0, that's not possible (gives error)
logvariations_computation(prices = c(100, 101, 99, 100), P=1, overlap=0)
}
}
