% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumvalues.R
\name{sumvalues}
\alias{sumvalues}
\title{Compute number of times returns (or values) are inside a box}
\usage{
sumvalues(rets, lims)
}
\arguments{
\item{rets}{A numeric vector with returns (or values) to be computed the Fisher Information.}

\item{lims}{A vector with the interval limits of width w from minimum to maximum value.}
}
\value{
A vector of the same length than lims with numbers of values inside each interval.
}
\description{
This is a function required to operate the Fisher Information function. It
computes the number of times that a values appears in each intervals.
}
\examples{
# Compute the intervals for a sequence of white noise
white_noise <- rnorm(50)
n <- 10 # number of intervals in which we split all the values of white noise
w <- diff(range(white_noise)) / n
lims <- seq(min(white_noise), max(white_noise), w) # we compute the interval limits
sumvalues(rets = white_noise, lims = lims)

}
